+++ FLEA/FLEA/Db/Driver/AbstractDriver.php
@@ -387,7 +387,7 @@ abstract class AbstractDriver
      */
     public function affectedRows(): int
     {
-        return $this->HAS_AFFECTED_ROWS ? $this->_affectedRows() : false;
+        return $this->HAS_AFFECTED_ROWS ? $this->_affectedRows() : 0;
     }
 
     /**
@@ -663,6 +663,7 @@ abstract class AbstractDriver
             $this->execute(\FLEA\Db\SqlStatement::create("SAVEPOINT {$savepoint}"));
             array_push($this->_savepointStack, $savepoint);
         }
+        return true;
     }
 
     /**
@@ -686,6 +687,7 @@ abstract class AbstractDriver
         } else {
             $this->_completeTrans($commitOnNoErrors);
         }
+        return true;
     }
 
     /**
@@ -800,7 +802,7 @@ abstract class AbstractDriver
      *
      * @return array
      */
-    public function getPlaceholderPair(array &$inputarr, $fields = null): string
+    public function getPlaceholderPair(array &$inputarr, $fields = null): array
     {
         $pairs = [];
         $values = [];